# `slack-19.0`: optimize `sqlutils.ToSqlite3Dialect`, part 2