--- a/go/vt/external/golib/sqlutils/sqlite_dialect.go
+++ b/go/vt/external/golib/sqlutils/sqlite_dialect.go
@@ -122,10 +122,10 @@ func ToSqlite3Insert(statement string) string {
 
 // ToSqlite3Dialect converts a statement to sqlite3 dialect. The statement
 // is checked in this order:
-//    1. If an insert/replace, convert with ToSqlite3Insert.
-//    2. If a create table, convert with IsCreateTable.
-//    3. If an alter table, convert with IsAlterTable.
-//    4. As fallback, return the statement with sqlite3GeneralConversions applied.
+//  1. If an insert/replace, convert with ToSqlite3Insert.
+//  2. If a create table, convert with IsCreateTable.
+//  3. If an alter table, convert with IsAlterTable.
+//  4. As fallback, return the statement with sqlite3GeneralConversions applied.
 func ToSqlite3Dialect(statement string) (translated string) {
 	if IsInsert(statement) {
 		return ToSqlite3Insert(statement)